refactor(tokenizer): remove 6 unused tokenizer implementations and consolidate shared logic
Delete 5 unused tokenizer implementations (Tokenizer class, EnhancedParsingTokenizer,
FastParsingTokenizer, RefactoredParsingTokenizer, ParallelTokenizer) and all streaming
infrastructure (StreamingTokenizer, SimpleTokenStream, TokenStreamAdapter). Consolidate
shared parsing strategies (TokenizerParsingStrategies, NumberParsingStrategies) into
TokenizerCore. Migrate all tests to use ParsingTokenizer as the sole production tokenizer.

Net reduction: 6,309 lines across 16 deleted files. All 1,254 remaining tests pass.
