--- Sources/FeLangCore/Tokenizer/RefactoredParsingTokenizer.swift
@@ -1,224 +1,0 @@
-import Foundation
-
-/// **DEMONSTRATION**: Refactored ParsingTokenizer using SharedTokenizerImplementation
-/// 
-/// This is an example of how the existing ParsingTokenizer can be refactored to use
-/// SharedTokenizerImplementation to eliminate code duplication while maintaining
-/// exactly the same functionality and performance characteristics.
-/// 
-/// **Code Reduction**: ~300 lines → ~150 lines (50% reduction)
-/// **Duplication Eliminated**: parseKeyword, parseOperator, parseDelimiter, parseNumber, 
-///                            parseHexadecimalNumber, parseBinaryNumber, parseOctalNumber, parseDecimalNumber
-/// **Benefits**: 
-/// - Consistent behavior across all tokenizers
-/// - Single point of maintenance for parsing logic
-/// - Easier testing (test shared implementation once vs. multiple copies)
-/// - Reduced binary size and compilation time
-public struct RefactoredParsingTokenizer: Sendable {
-
-    // MARK: - Initialization
-
-    public init() {}
-
-    // MARK: - Tokenizer Implementation
-
-    public func tokenize(_ input: String) throws -> [Token] {
-        var tokens: [Token] = []
-        var index = input.startIndex
-        let startIndex = index
-
-        while index < input.endIndex {
-            let position = TokenizerUtilities.sourcePosition(from: input, startIndex: startIndex, currentIndex: index)
-
-            // Skip whitespace and newlines (same as original ParsingTokenizer)
-            if input[index].isWhitespace {
-                index = input.index(after: index)
-                continue
-            }
-
-            // Try to parse a token using shared implementation
-            let beforeIndex = index
-            if let token = try parseNextToken(from: input, at: &index, startIndex: startIndex) {
-                let tokenWithPosition = Token(
-                    type: token.type,
-                    lexeme: token.lexeme,
-                    position: position
-                )
-                tokens.append(tokenWithPosition)
-            } else {
-                // Check if index moved (could be a comment that was skipped)
-                if index > beforeIndex {
-                    continue // Comment was skipped, continue to next iteration
-                }
-
-                // If we can't parse a token, it's an unexpected character
-                guard let scalar = String(input[index]).unicodeScalars.first else {
-                    index = input.index(after: index)
-                    continue
-                }
-                throw TokenizerError.unexpectedCharacter(scalar, position)
-            }
-
-            // Safety check to prevent infinite loops
-            if index == beforeIndex {
-                guard let scalar = String(input[index]).unicodeScalars.first else {
-                    index = input.index(after: index)
-                    continue
-                }
-                throw TokenizerError.unexpectedCharacter(scalar, position)
-            }
-        }
-
-        // Add EOF token
-        let finalPosition = TokenizerUtilities.sourcePosition(from: input, startIndex: startIndex, currentIndex: index)
-        tokens.append(Token(type: .eof, lexeme: "", position: finalPosition))
-
-        return tokens
-    }
-
-    // MARK: - Private Token Parsing Logic (using SharedTokenizerImplementation)
-
-    private func parseNextToken(from input: String, at index: inout String.Index, startIndex: String.Index) throws -> SharedTokenizerImplementation.TokenData? {
-        // Try to parse comments first (skip them, don't return tokens) - same as original
-        if try parseComment(from: input, at: &index, startIndex: startIndex) != nil {
-            return nil // Comments are skipped
-        }
-
-        // Try to parse keywords using shared implementation
-        if let token = SharedTokenizerImplementation.parseKeyword(from: input, at: &index) {
-            return token
-        }
-
-        // Try to parse operators using shared implementation  
-        if let token = SharedTokenizerImplementation.parseOperator(from: input, at: &index) {
-            return token
-        }
-
-        // Try to parse numbers using shared implementation (including leading-dot decimals)
-        if let token = SharedTokenizerImplementation.parseNumber(from: input, at: &index) {
-            return token
-        }
-
-        // Try to parse delimiters using shared implementation
-        if let token = SharedTokenizerImplementation.parseDelimiter(from: input, at: &index) {
-            return token
-        }
-
-        // Try to parse strings using shared implementation
-        if let token = try parseString(from: input, at: &index) {
-            return token
-        }
-
-        // Try to parse identifiers using shared implementation
-        if let token = SharedTokenizerImplementation.parseIdentifier(from: input, at: &index) {
-            return token
-        }
-
-        return nil
-    }
-
-    // MARK: - Private Helpers (Tokenizer-Specific Logic)
-
-    /// Parses comments - this remains tokenizer-specific since different tokenizers
-    /// may have different comment handling requirements
-    private func parseComment(from input: String, at index: inout String.Index, startIndex: String.Index) throws -> SharedTokenizerImplementation.TokenData? {
-        guard index < input.endIndex else { return nil }
-
-        // Single line comment
-        if TokenizerUtilities.matchString("//", in: input, at: index) {
-            let start = index
-            index = input.index(index, offsetBy: 2)
-
-            // Read until newline or end
-            while index < input.endIndex && input[index] != "\n" {
-                index = input.index(after: index)
-            }
-
-            let lexeme = String(input[start..<index])
-            return SharedTokenizerImplementation.TokenData(type: .comment, lexeme: lexeme)
-        }
-
-        // Multi-line comment
-        if TokenizerUtilities.matchString("/*", in: input, at: index) {
-            let commentStart = index
-            let position = TokenizerUtilities.sourcePosition(from: input, startIndex: startIndex, currentIndex: index)
-            index = input.index(index, offsetBy: 2)
-
-            var foundTerminator = false
-            // Read until */
-            while index < input.endIndex {
-                if TokenizerUtilities.matchString("*/", in: input, at: index) {
-                    index = input.index(index, offsetBy: 2)
-                    foundTerminator = true
-                    break
-                }
-                index = input.index(after: index)
-            }
-
-            // Check if comment was properly terminated
-            if !foundTerminator {
-                throw TokenizerError.unterminatedComment(position)
-            }
-
-            let lexeme = String(input[commentStart..<index])
-            return SharedTokenizerImplementation.TokenData(type: .comment, lexeme: lexeme)
-        }
-
-        return nil
-    }
-
-    /// Parses string literals - demonstrates how string parsing can use shared implementation
-    private func parseString(from input: String, at index: inout String.Index) throws -> SharedTokenizerImplementation.TokenData? {
-        guard index < input.endIndex else { return nil }
-
-        if input[index] == "'" {
-            switch SharedTokenizerImplementation.parseStringLiteral(from: input, at: &index, quoteChar: "'") {
-            case .success(let token):
-                return token
-            case .failure(let error):
-                throw error
-            }
-        }
-
-        return nil
-    }
-}
-
-// MARK: - Performance Comparison Documentation
-
-/*
- ## Performance Comparison: Original vs Refactored
- 
- ### Original ParsingTokenizer:
- - Lines of Code: ~500
- - Duplicated Methods: 8 major parsing methods
- - Binary Size Impact: Each duplicated method adds ~1-2KB
- - Maintenance Burden: Changes must be made in 4+ places
- 
- ### Refactored ParsingTokenizer:
- - Lines of Code: ~180 (64% reduction)
- - Duplicated Methods: 0 (all use SharedTokenizerImplementation)
- - Binary Size Impact: Shared methods compiled once
- - Maintenance Burden: Changes made in one place
- 
- ### Runtime Performance:
- - **Identical**: Shared implementation uses same algorithms
- - **Memory**: Slightly better due to less code duplication
- - **CPU**: No difference in parsing speed
- - **Compilation**: Faster due to less duplicated code
- 
- ### Migration Path:
- 1. Replace parseKeyword → SharedTokenizerImplementation.parseKeyword
- 2. Replace parseOperator → SharedTokenizerImplementation.parseOperator  
- 3. Replace parseDelimiter → SharedTokenizerImplementation.parseDelimiter
- 4. Replace parseNumber → SharedTokenizerImplementation.parseNumber
- 5. Keep tokenizer-specific logic (comments, options) unchanged
- 6. Run tests to ensure identical behavior
- 7. Gradually migrate other tokenizers using same pattern
- 
- ### Risk Assessment:
- - **Low Risk**: Shared implementation uses exact same algorithms
- - **High Confidence**: Can be done incrementally, method by method
- - **Easy Rollback**: Original methods can be restored if needed
- - **Test Coverage**: Existing tests validate behavior is unchanged
- */
