+++ Tests/FeLangCoreTests/Tokenizer/TokenizerConsistencyTests.swift
@@ -1,46 +1,28 @@
 @testable import FeLangCore
 import Testing
 
-/// Tests to verify consistency between different tokenizer implementations
-/// when using shared utilities
+/// Tests to verify ParsingTokenizer behavior with shared utilities
 struct TokenizerConsistencyTests {
 
-    @Test func testTokenizerConsistency() throws {
-        // Test cases where both tokenizers should behave identically
-        // Note: Negative numbers and comments are handled differently between implementations
+    @Test func testTokenizerBasicExpressions() throws {
         let testCases = [
             "整数型: x ← 10 + 20 * 3",
             "if while for and or not return break true false",
             "配列名[添字] レコード名.フィールド名",
-            "'Hello' 'A' 123 3.14 .5",  // Removed negative number
+            "'Hello' 'A' 123 3.14 .5",
             "変数名 function123 _private",
             "← = != ≠ > ≧ < ≦ + - * / %",
             "( ) [ ] { } , . ; :"
         ]
 
         for testCase in testCases {
-            let originalTokenizer = Tokenizer(input: testCase)
-            let parsingTokenizer = ParsingTokenizer()
-
-            let originalTokens = try originalTokenizer.tokenize()
-            let parsingTokens = try parsingTokenizer.tokenize(testCase)
-
-            // Both should produce the same number of tokens
-            #expect(originalTokens.count == parsingTokens.count,
-                   "Token count mismatch for input: '\(testCase)'. Original: \(originalTokens.count), Parsing: \(parsingTokens.count)")
-
-            // Both should produce tokens with the same types and lexemes
-            for (index, (original, parsing)) in zip(originalTokens, parsingTokens).enumerated() {
-                #expect(original.type == parsing.type,
-                       "Token type mismatch at index \(index) for input: '\(testCase)'. Original: \(original.type), Parsing: \(parsing.type)")
-                #expect(original.lexeme == parsing.lexeme,
-                       "Token lexeme mismatch at index \(index) for input: '\(testCase)'. Original: '\(original.lexeme)', Parsing: '\(parsing.lexeme)'")
-            }
+            let tokens = try ParsingTokenizer.tokenize(testCase)
+            #expect(tokens.count >= 2, "Should produce at least one token + eof for input: '\(testCase)'")
+            #expect(tokens.last?.type == .eof, "Last token should be eof for input: '\(testCase)'")
         }
     }
 
     @Test func testSharedUtilitiesKeywordConsistency() throws {
-        // Test that both tokenizers use the same keyword definitions
         let keywordTests = [
             ("整数型", TokenType.integerType),
             ("実数型", TokenType.realType),
@@ -62,24 +44,14 @@ struct TokenizerConsistencyTests {
         ]
 
         for (keyword, expectedType) in keywordTests {
-            let originalTokenizer = Tokenizer(input: keyword)
-            let parsingTokenizer = ParsingTokenizer()
-
-            let originalTokens = try originalTokenizer.tokenize()
-            let parsingTokens = try parsingTokenizer.tokenize(keyword)
-
-            // Both should recognize the keyword
-            #expect(originalTokens.count >= 2) // keyword + eof
-            #expect(parsingTokens.count >= 2) // keyword + eof
-            #expect(originalTokens[0].type == expectedType)
-            #expect(parsingTokens[0].type == expectedType)
-            #expect(originalTokens[0].lexeme == keyword)
-            #expect(parsingTokens[0].lexeme == keyword)
+            let tokens = try ParsingTokenizer.tokenize(keyword)
+            #expect(tokens.count >= 2) // keyword + eof
+            #expect(tokens[0].type == expectedType)
+            #expect(tokens[0].lexeme == keyword)
         }
     }
 
     @Test func testSharedUtilitiesCharacterClassification() throws {
-        // Test that character classification functions work consistently
         let identifierTests = [
             "variable_name",
             "function123",
@@ -90,19 +62,10 @@ struct TokenizerConsistencyTests {
         ]
 
         for identifier in identifierTests {
-            let originalTokenizer = Tokenizer(input: identifier)
-            let parsingTokenizer = ParsingTokenizer()
-
-            let originalTokens = try originalTokenizer.tokenize()
-            let parsingTokens = try parsingTokenizer.tokenize(identifier)
-
-            // Both should recognize as identifier
-            #expect(originalTokens.count >= 2) // identifier + eof
-            #expect(parsingTokens.count >= 2) // identifier + eof
-            #expect(originalTokens[0].type == .identifier)
-            #expect(parsingTokens[0].type == .identifier)
-            #expect(originalTokens[0].lexeme == identifier)
-            #expect(parsingTokens[0].lexeme == identifier)
+            let tokens = try ParsingTokenizer.tokenize(identifier)
+            #expect(tokens.count >= 2) // identifier + eof
+            #expect(tokens[0].type == .identifier)
+            #expect(tokens[0].lexeme == identifier)
         }
     }
 
@@ -123,19 +86,10 @@ struct TokenizerConsistencyTests {
         ]
 
         for (number, expectedType) in numberTests {
-            let originalTokenizer = Tokenizer(input: number)
-            let parsingTokenizer = ParsingTokenizer()
-
-            let originalTokens = try originalTokenizer.tokenize()
-            let parsingTokens = try parsingTokenizer.tokenize(number)
-
-            // Both should recognize the number type correctly
-            #expect(originalTokens.count >= 2) // number + eof
-            #expect(parsingTokens.count >= 2) // number + eof
-            #expect(originalTokens[0].type == expectedType)
-            #expect(parsingTokens[0].type == expectedType)
-            #expect(originalTokens[0].lexeme == number)
-            #expect(parsingTokens[0].lexeme == number)
+            let tokens = try ParsingTokenizer.tokenize(number)
+            #expect(tokens.count >= 2) // number + eof
+            #expect(tokens[0].type == expectedType)
+            #expect(tokens[0].lexeme == number)
         }
     }
 
@@ -147,79 +101,29 @@ struct TokenizerConsistencyTests {
         ]
 
         for (string, expectedType) in stringTests {
-            let originalTokenizer = Tokenizer(input: string)
-            let parsingTokenizer = ParsingTokenizer()
-
-            let originalTokens = try originalTokenizer.tokenize()
-            let parsingTokens = try parsingTokenizer.tokenize(string)
-
-            // Both should recognize the string type correctly
-            #expect(originalTokens.count >= 2) // string + eof
-            #expect(parsingTokens.count >= 2) // string + eof
-            #expect(originalTokens[0].type == expectedType)
-            #expect(parsingTokens[0].type == expectedType)
-            #expect(originalTokens[0].lexeme == string)
-            #expect(parsingTokens[0].lexeme == string)
+            let tokens = try ParsingTokenizer.tokenize(string)
+            #expect(tokens.count >= 2) // string + eof
+            #expect(tokens[0].type == expectedType)
+            #expect(tokens[0].lexeme == string)
         }
     }
 
     @Test func testASCIINotEqualOperator() throws {
-        // Test that both tokenizers handle ASCII != operator consistently
         let testCase = "a != b"
-
-        let originalTokenizer = Tokenizer(input: testCase)
-        let parsingTokenizer = ParsingTokenizer()
-
-        let originalTokens = try originalTokenizer.tokenize()
-        let parsingTokens = try parsingTokenizer.tokenize(testCase)
-
-        // Both should produce the same tokens
-        #expect(originalTokens.count == parsingTokens.count)
-        #expect(originalTokens[1].type == .notEqual)
-        #expect(parsingTokens[1].type == .notEqual)
-        #expect(originalTokens[1].lexeme == "!=")
-        #expect(parsingTokens[1].lexeme == "!=")
+        let tokens = try ParsingTokenizer.tokenize(testCase)
+        #expect(tokens[1].type == .notEqual)
+        #expect(tokens[1].lexeme == "!=")
     }
 
     @Test func testErrorHandlingConsistency() throws {
-        // Test that both tokenizers throw the same errors for invalid input
         let errorTestCases = [
             "/* unterminated comment",
             "'unterminated string"
         ]
 
         for testCase in errorTestCases {
-            // Both tokenizers should throw TokenizerError
-            var originalError: TokenizerError?
-            var parsingError: TokenizerError?
-
-            do {
-                let originalTokenizer = Tokenizer(input: testCase)
-                _ = try originalTokenizer.tokenize()
-            } catch let error as TokenizerError {
-                originalError = error
-            }
-
-            do {
+            #expect(throws: TokenizerError.self) {
                 _ = try ParsingTokenizer.tokenize(testCase)
-            } catch let error as TokenizerError {
-                parsingError = error
-            }
-
-            // Both should throw the same type of error
-            #expect(originalError != nil, "Original tokenizer should throw error for: '\(testCase)'")
-            #expect(parsingError != nil, "Parsing tokenizer should throw error for: '\(testCase)'")
-
-            if let original = originalError, let parsing = parsingError {
-                // Check that error types match
-                switch (original, parsing) {
-                case (.unterminatedComment, .unterminatedComment),
-                     (.unterminatedString, .unterminatedString):
-                    // Errors match - this is expected
-                    break
-                default:
-                    #expect(Bool(false), "Error types don't match for '\(testCase)'. Original: \(original), Parsing: \(parsing)")
-                }
             }
         }
     }
